+++ DialMock.CadModel/Geometry/CadPoint2.cs
@@ -1,0 +1,3 @@
+namespace DialMock.CadModel.Geometry;
+
+public readonly record struct CadPoint2(double X, double Y);