+++ DialMock.CadModel/Model/CadEntity.cs
@@ -1,0 +1,6 @@
+namespace DialMock.CadModel.Model;
+
+public abstract record CadEntity
+{
+    public string LayerName { get; init; } = "0";
+}