+++ DialMock.CadModel/Model/CadLayer.cs
@@ -1,0 +1,6 @@
+namespace DialMock.CadModel.Model;
+
+public sealed record CadLayer
+{
+    public required string Name { get; init; }
+}